Solver

TimetableSolver.VarRefOrExpr — Type

VarRefOrExpr

Represents either a model variable reference or expression reference.

Alias for union of VariableRef and AffExpr.

TimetableSolver.VariableData — Type

VariableData

Immutable struct to store data related to divisions, variables and values, and their mappings from ints to strings, because the solver only works on ints.

Fields:

divvarval_maps::OrderedDict{Symbol,OrderedDict}: Keys are a symbol :subject or :teacher. Values are map of ints (div) to 2-tuple of vecs of ints (var, val).

divvarval_mapstrs::OrderedDict{Symbol,OrderedDict}: Same as div_var_val_map but with strings instead of ints.

alldivs::Vector{Int}: Vector of all divs as ints.

allvars::Vector{Int}: Vector of all vars as ints. (no :subject :teacher bifurcation)

allvals::Vector{Int}: Vector of all vals as ints. (no :subject :teacher bifurcation)

divmap::OrderedDict{Int,String}: Map from div ints to strings.

varmap::OrderedDict{Int,String}: Map from variable ints to strings.

valmap::OrderedDict{Int,String}: Map from values ints to strings.

inversedivmap::OrderedDict{String,Int}: Inverted divmap.

inversevarmap::OrderedDict{String,Int}: Inverted varmap.

inversevalmap::OrderedDict{String,Int}: Inverted valmap.

schedule::Schedule: Schedule type instance.

TimetableSolver.VariableData — Method

VariableData(schedule::Schedule)::VariableData

Return VariableData instance with all the fields filled based on the schedule.

TimetableSolver.get_divvarval_mapstrs — Method

get_divvarval_mapstrs(schedule::Schedule)::OrderedDict{Symbol,OrderedDict}

Return OrderedDict mapping symbol (either :subject or :teacher), to an OrderedDict.

The first-order values (OrderedDict), are mappings from division string, to a 2-tuple of variables and values. "variables" and "values" are vectors of variable strings and value strings.

TimetableSolver.get_model — Function

get_model(vardata::VariableData, timeout::Float64=Inf, all_solutions::Bool=false)::Tuple{Model,OrderedDict}

Return a JuMP model created using vardata, and an OrderedDict of the variables defined in the model.

Arguments

vardata::VariableData: VariableData object representing the variables and constraints to be added to the model.

timeout::Float64=Inf: Float representing the maximum time to run the model.

all_solutions::Bool=false: Bool representing whether to return all solutions or just the first.

Notes

Create a model with the correct variables, and adds constraints to the model.

The variables defined are anonymous, so they are stored in an OrderedDict mapped to their int representation.

TimetableSolver.define_variables! — Function

define_variables!(m::Model, vardata::VariableData)::OrderedDict{Int,VarRefOrExpr}

Define variables in the model m using the vardata object. Return an OrderedDict mapping the int representation of the variable to the variable reference/expression reference.

Notes

Modify the model m in place.

See VarRefOrExpr.

TimetableSolver.define_subjectconstraints! — Function

define_subjectconstraints!(m::Model, modelvars::OrderedDict, vardata::VariableData)::Nothing

Define subject constraints in the model m using the vardata object and the model variables modelvars.

Notes

Modify the model m in place.

TimetableSolver.define_subjectteacherconstraints! — Function

define_subjectteacherconstraints!(m::Model, modelvars::OrderedDict, vardata::VariableData)::Nothing

Define subject-teacher constraints in the model m using the vardata object and the model variables modelvars.

Notes

Modify the model m in place.

TimetableSolver.define_teacherconstraints! — Function

define_teacherconstraints!(m::Model, modelvars::OrderedDict, vardata::VariableData)::Nothing

Define teacher constraints in the model m using the vardata object and the model variables modelvars.

Notes

Modify the model m in place.

TimetableSolver.define_constraints! — Function

define_constraints!(m::Model, modelvars::OrderedDict, vardata::VariableData)::Nothing

Define all constraints in the model m using the vardata object and the model variables modelvars.

Notes

Modify the model m in place.

Call define_subjectconstraints!, define_subjectteacherconstraints! and define_teacherconstraints! internally, in the respective order.

TimetableSolver.get_solution — Function

get_solution(m::Model, modelvars::OrderedDict, vardata::VariableData)::Tuple{OrderedDict{Int,Int},MOI.TerminationStatusCode}

Return a solution and status code after solving the given model m and using model variables modelvars and variable data vardata to access the solution values.

Notes

Model is optimally solved only if the status code is MOI.OPTIMAL.

Solution is a mapping from variable to its value (both as ints).

TimetableSolver.convertsolution — Function

convertsolution(rawsolution::OrderedDict{Int,Int}, vardata::VariableData)::OrderedDict{String,String}

Return a solution in the form of a mapping from variable to its value (both as strings).

Notes

Convert an Int=>Int solution to a String=>String one using the inverse* mappings defined in vardata.

TimetableSolver.applysolution! — Function

applysolution!(schedule::Schedule, solution::OrderedDict{String,String})::Nothing

Apply a solution (of strings) to the given schedule.

Notes

Modify the schedule in place.

TimetableSolver.solve! — Function

solve!(schedule::Schedule, timeout::Float64 = Inf, all_solutions::Bool = false)::Tuple{OrderedDict{String,String},MOI.TerminationStatusCode}

Solve a given schedule in place and return a String=>String solution and status code.

Arguments

schedule::Schedule: Schedule object representing the problem.

timeout::Float64=Inf: Float representing the maximum time to run the model.

all_solutions::Bool=false: Bool representing whether to return all solutions or just the first.

Notes

Schedule is optimally solved only if the status code is MOI.OPTIMAL.